move block improvement context factory to api
