+++ src/Nethermind/Nethermind.Api/Extensions/IConsensusWrapperPlugin.cs
@@ -1,7 +1,6 @@
 // SPDX-FileCopyrightText: 2022 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
-using System.Threading.Tasks;
 using Nethermind.Consensus;
 using Nethermind.Consensus.Transactions;
