+++ src/Nethermind/Nethermind.Shutter/ShutterPlugin.cs
@@ -53,18 +53,15 @@ namespace Nethermind.Shutter
 
         public Task InitRpcModules()
         {
-            // todo: how to do this?
-            // IBlockImprovementContextFactory? blockImprovementContextFactory = null;
-            // if (_shutterConfig!.Enabled)
-            // {
-            //     blockImprovementContextFactory = new ShutterBlockImprovementContextFactory(
-            //         _api.BlockProducer!,
-            //         _shutterTxSource!,
-            //         _shutterConfig,
-            //         _api.SpecProvider!,
-            //         _api.LogManager);
-            // }
-            // base.InitRpcModulesInternal(blockImprovementContextFactory, false);
+            if (Enabled)
+            {
+                _api!.BlockImprovementContextFactory = new ShutterBlockImprovementContextFactory(
+                    _api.BlockProducer!,
+                    _shutterTxSource!,
+                    _shutterConfig!,
+                    _api.SpecProvider!,
+                    _api.LogManager);
+            }
             return Task.CompletedTask;
         }
